refactor: 버스 도메인 모델 구조 개선 #355
- BusRealTimeArrival을 BusRealTimeArrivals로 변경하여 복수형으로 일관성 확보
- BusArrivalInfo 도메인 모델 추가로 도착 정보 캡슐화
- BusArrival, BusSchedule 모델 개선
