+++ src/main/kotlin/com/deepromeet/atcha/transit/domain/bus/BusArrivalInfo.kt
@@ -1,0 +1,6 @@
+package com.deepromeet.atcha.transit.domain.bus
+
+data class BusArrivalInfo(
+    val schedule: BusSchedule,
+    val realTimeArrival: BusRealTimeArrivals
+)
